[MODGOBI-152] - Code Refactoring
