+++ src/test/java/org/folio/gobi/MappingHelperTest.java
@@ -18,10 +18,10 @@ import static org.junit.Assert.fail;
 import java.io.IOException;
 import java.io.InputStream;
 import java.nio.charset.Charset;
+import java.util.EnumMap;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.stream.Collectors;
 
 import org.apache.commons.io.IOUtils;
 import org.apache.logging.log4j.LogManager;
@@ -37,24 +37,29 @@ import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
 import mockit.Mock;
 import mockit.MockUp;
+import org.mockito.Mockito;
 
 public class MappingHelperTest {
 
-  private static final Logger logger = LogManager.getLogger(MappingHelper.class);
-  private static final String LISTED_PRINT_PATH = "MappingHelper/ListedPrintMonograph.json";
-  private static final String UNLISTED_PRINT_PATH = "MappingHelper/UnlistedPrintMonograph.json";
+  private final Logger logger = LogManager.getLogger(MappingHelper.class);
+  private final String LISTED_PRINT_PATH = "MappingHelper/ListedPrintMonograph.json";
+  private final String UNLISTED_PRINT_PATH = "MappingHelper/UnlistedPrintMonograph.json";
 
-  private static final String EXPECTEDORDERTYPE1 = "ListedPrintMonograph";
-  private static final String EXPECTEDORDERTYPE2 = "UnlistedPrintMonograph";
+  private final String EXPECTEDORDERTYPE1 = "ListedPrintMonograph";
+  private final String EXPECTEDORDERTYPE2 = "UnlistedPrintMonograph";
 
-  private static String actualListedPrintJson = null;
-  private static String actualUnlistedPrintJson = null;
-  private static OrderMappings actualUnlistedPrintMappings = null;
-  private static OrderMappings actuallistedPrintMappings = null;
-  private static Mapping mappingAccessProvider = null;
-  private static Mapping mappingExpenseClass = null;
-  private static Mapping mappingCombinatorNotExist = null;
-  private static Mapping mapping2 = null;
+  private String actualListedPrintJson = null;
+  private String actualUnlistedPrintJson = null;
+  private OrderMappings actualUnlistedPrintMappings = null;
+  private OrderMappings actuallistedPrintMappings = null;
+  private Mapping mappingAccessProvider = null;
+  private Mapping mappingExpenseClass = null;
+  private Mapping mappingCombinatorNotExist = null;
+  private Mapping mapping2 = null;
+  private Map<Mapping.Field, org.folio.gobi.DataSourceResolver> fieldDataSourceMapping = new EnumMap<>(Mapping.Field.class);
+  private LookupService lookupService = Mockito.mock(LookupService.class);
+  private FieldMappingTranslatorResolver fieldMappingTranslatorResolver = new FieldMappingTranslatorResolver(lookupService);
+  private MappingHelper mappingHelper = new MappingHelper(fieldMappingTranslatorResolver);
 
   @Before
   public void setUp() {
@@ -85,20 +90,21 @@ public class MappingHelperTest {
       .put("dataSource",
         new JsonObject().put("from", "//LocalData[Description='LocalData4']/Value").put("translation", "lookupConfigAddress"));;
 
-    JsonObject linkedPackageId = new JsonObject().put("field", LINKED_PACKAGE.value())
-      .put("dataSource",
-        new JsonObject().put("from", "//LocalData[Description='LocalData5']/Value").put("translation", "lookupLinkedPackage"));;
-
-
     JsonObject expenseClass = new JsonObject().put("field", EXPENSE_CLASS.value())
       .put("dataSource",
-        new JsonObject().put("from", "//LocalData[Description='LocalData6']/Value").put("translation", "lookupExpenseClassId"));;
+        new JsonObject().put("from", "//LocalData[Description='LocalData5']/Value").put("translation", "lookupExpenseClassId"));;
+
+    JsonObject linkedPackageId = new JsonObject().put("field", LINKED_PACKAGE.value())
+      .put("dataSource",
+        new JsonObject().put("from", "//LocalData[Description='LocalData6']/Value").put("translation", "lookupLinkedPackage"));;
+
 
     JsonObject combinatorTitle = new JsonObject().put("field", TITLE.value())
       .put("dataSource",
-        new JsonObject().put("from", "//datafield[@tag='245']/*").put("combinator", "concat"));;
+        new JsonObject().put("from", "//datafield[@tag='245']/*").put("combinator", "concat"));
+
     expectedListedPrintMonographJsonObj.put("mappings",
-      new JsonArray().add(accessProvider).add(expenseClass).add(combinatorTitle).add(linkedPackageId)
+      new JsonArray().add(accessProvider).add(combinatorTitle).add(linkedPackageId).add(expenseClass)
                         .add(shipToId).add(billToId).add(suffixId).add(prefixId));
 
     String expectedListedPrintMonographJson = expectedListedPrintMonographJsonObj.toString();
@@ -135,7 +141,7 @@ public class MappingHelperTest {
     List<Mapping> mappingsList1 = actuallistedPrintMappings.getMappings();
     List<Mapping> mappingsList2 = actualUnlistedPrintMappings.getMappings();
     mappingAccessProvider = mappingsList1.get(0);
-    mappingExpenseClass = mappingsList1.get(1);
+    mappingExpenseClass = mappingsList1.get(3);
     mappingCombinatorNotExist = mappingsList1.get(2);
     mappingCombinatorNotExist.getDataSource().setCombinator(null);
 
@@ -177,7 +183,7 @@ public class MappingHelperTest {
     logger.info(
         "Begin: Testing for default mapping when it returns a DataSource mapping for OrderType ListedPrintMonograph");
     Map<Mapping.Field, org.folio.gobi.DataSourceResolver> fieldDataSourceMapping1 = new LinkedHashMap<>();
-    org.folio.gobi.DataSourceResolver dataSource = MappingHelper.getDS(mappingAccessProvider, fieldDataSourceMapping1, null);
+    org.folio.gobi.DataSourceResolver dataSource = mappingHelper.getDS(mappingAccessProvider, fieldDataSourceMapping1, null);
     assertEquals("//PurchaseOption/VendorPOCode", dataSource.from);
     assertFalse(dataSource.translateDefValue);
   }
@@ -187,7 +193,7 @@ public class MappingHelperTest {
     logger.info(
         "Begin: Testing for default mapping when it returns a DataSource mapping for OrderType UnlistedPrintMonograph");
     Map<Mapping.Field, org.folio.gobi.DataSourceResolver> fieldDataSourceMapping2 = new LinkedHashMap<>();
-    org.folio.gobi.DataSourceResolver dataSource = MappingHelper.getDS(mapping2, fieldDataSourceMapping2, null);
+    org.folio.gobi.DataSourceResolver dataSource = mappingHelper.getDS(mapping2, fieldDataSourceMapping2, null);
     assertEquals("//datafield[@tag='020']/subfield[@code='a']", dataSource.from);
     assertFalse(dataSource.translateDefValue);
     assertNotNull(dataSource.combinator);
@@ -213,24 +219,30 @@ public class MappingHelperTest {
   @Test
   public void testShouldSuccessMapDiffLookups() {
     Map<Mapping.Field, List<Mapping>> mappings = actuallistedPrintMappings.getMappings().stream().collect(groupingBy(Mapping::getField));
-    org.folio.gobi.DataSourceResolver dataSourcePrefix = MappingHelper.getDS(mappings.get(PREFIX).get(0), new LinkedHashMap<>(), null);
+    org.folio.gobi.DataSourceResolver dataSourcePrefix = mappingHelper.getDS(mappings.get(PREFIX).get(0), fieldDataSourceMapping, null);
 
     assertEquals("//LocalData[Description='LocalData1']/Value", dataSourcePrefix.from);
-    org.folio.gobi.DataSourceResolver dataSourceSuffix = MappingHelper.getDS(mappings.get(SUFFIX).get(0), new LinkedHashMap<>(), null);
+    assertNotNull(dataSourcePrefix.translation);
+    org.folio.gobi.DataSourceResolver dataSourceSuffix = mappingHelper.getDS(mappings.get(SUFFIX).get(0), fieldDataSourceMapping, null);
     assertEquals("//LocalData[Description='LocalData2']/Value", dataSourceSuffix.from);
-    org.folio.gobi.DataSourceResolver dataSourceBillTo = MappingHelper.getDS(mappings.get(BILL_TO).get(0), new LinkedHashMap<>(), null);
-     assertEquals("//LocalData[Description='LocalData3']/Value", dataSourceBillTo.from);
-    org.folio.gobi.DataSourceResolver dataSourceShipTo = MappingHelper.getDS(mappings.get(SHIP_TO).get(0), new LinkedHashMap<>(), null);
+    assertNotNull(dataSourceSuffix.translation);
+    org.folio.gobi.DataSourceResolver dataSourceBillTo = mappingHelper.getDS(mappings.get(BILL_TO).get(0),fieldDataSourceMapping, null);
+    assertEquals("//LocalData[Description='LocalData3']/Value", dataSourceBillTo.from);
+    assertNotNull(dataSourceBillTo.translation);
+    org.folio.gobi.DataSourceResolver dataSourceShipTo = mappingHelper.getDS(mappings.get(SHIP_TO).get(0), fieldDataSourceMapping, null);
     assertEquals("//LocalData[Description='LocalData4']/Value", dataSourceShipTo.from);
-    org.folio.gobi.DataSourceResolver dataSourceLinkedPackage = MappingHelper.getDS(mappings.get(LINKED_PACKAGE).get(0), new LinkedHashMap<>(), null);
-    assertEquals("//LocalData[Description='LocalData5']/Value", dataSourceLinkedPackage.from);
+    assertNotNull(dataSourceShipTo.translation);
+    org.folio.gobi.DataSourceResolver dataSourceLinkedPackage = mappingHelper.getDS(mappings.get(LINKED_PACKAGE).get(0), fieldDataSourceMapping, null);
+    assertNotNull(dataSourceLinkedPackage.translation);
+    assertEquals("//LocalData[Description='LocalData6']/Value", dataSourceLinkedPackage.from);
   }
 
   @Test
   public void testShouldSuccessMapExpenseClass() {
     Map<Mapping.Field, org.folio.gobi.DataSourceResolver> fieldDataSourceMapping = new LinkedHashMap<>();
-    org.folio.gobi.DataSourceResolver dataSource = MappingHelper.getDS(mappingExpenseClass, fieldDataSourceMapping, null);
+    org.folio.gobi.DataSourceResolver dataSource = mappingHelper.getDS(mappingExpenseClass, fieldDataSourceMapping, null);
     assertEquals("//LocalData[Description='LocalData5']/Value", dataSource.from);
+    assertNotNull(dataSource.translation);
     assertFalse(dataSource.translateDefValue);
   }
 }
